+++ app/Http/Controllers/CategoryController.php
@@ -6,6 +6,8 @@ use App\Models\Category;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
+
 
 class CategoryController extends Controller
 {
@@ -14,6 +16,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        // $this->authorize("viewAny",Category::class);
         $categories = Category::latest("id")->get();
         return view("category.index",compact("categories"));
     }
@@ -23,6 +26,9 @@ class CategoryController extends Controller
      */
     public function create()
     {
+
+        // $this->authorize("create",Category::class);
+
         return view("category.create");
     }
 
@@ -31,8 +37,12 @@ class CategoryController extends Controller
      */
     public function store(StoreCategoryRequest $request)
     {
+
+        // $this->authorize("viewAny",Category::class);
+
         Category::create([
             "title" => $request->title,
+            "slug" => Str::slug($request->title),
             "user_id" => Auth::id()
         ]);
 
@@ -52,6 +62,7 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        $this->authorize("update",$category);
         return view("category.edit",compact('category'));
     }
 
@@ -60,8 +71,12 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, Category $category)
     {
+
+        $this->authorize("update",$category);
+
         $category->update([
-            "title" => $request->title
+            "title" => $request->title,
+            "slug" => Str::slug($request->title)
         ]);
 
         return redirect()->route("category.index");
@@ -72,6 +87,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize("delete",$category);
+
         $category->delete();
         return redirect()->back();
     }
